+++ packages/binance_margin/lib/binance_margin.dart
@@ -1,0 +1,7 @@
+/// Support for Binance Margin API.
+library;
+
+export 'src/enums.dart';
+export 'src/models/account.dart';
+export 'src/models/history.dart';
+export 'src/rest/margin_client.dart';